+++ app/kumactl/cmd/inspect/inspect_dataplanes_test.go
@@ -27,12 +27,14 @@ import (
 )
 
 type testDataplaneOverviewClient struct {
-	receivedTags map[string]string
-	overviews    []*mesh_core.DataplaneOverviewResource
+	receivedTags    map[string]string
+	receivedGateway bool
+	overviews       []*mesh_core.DataplaneOverviewResource
 }
 
-func (c *testDataplaneOverviewClient) List(_ context.Context, _ string, tags map[string]string) (*mesh_core.DataplaneOverviewResourceList, error) {
+func (c *testDataplaneOverviewClient) List(_ context.Context, _ string, tags map[string]string, gateway bool) (*mesh_core.DataplaneOverviewResourceList, error) {
 	c.receivedTags = tags
+	c.receivedGateway = gateway
 	return &mesh_core.DataplaneOverviewResourceList{
 		Items: c.overviews,
 	}, nil
@@ -242,6 +244,22 @@ var _ = Describe("kumactl inspect dataplanes", func() {
 				// and
 				Expect(testClient.receivedTags).To(HaveKeyWithValue("service", "mobile"))
 				Expect(testClient.receivedTags).To(HaveKeyWithValue("version", "v1"))
+			})
+		})
+
+		Describe("kumactl inspect dataplanes --gateway", func() {
+			It("gateway should be passed to the client", func() {
+				// given
+				rootCmd.SetArgs([]string{
+					"--config-file", filepath.Join("..", "testdata", "sample-kumactl.config.yaml"),
+					"inspect", "dataplanes", "--gateway"})
+
+				// when
+				err := rootCmd.Execute()
+				// then
+				Expect(err).ToNot(HaveOccurred())
+				// and
+				Expect(testClient.receivedGateway).To(BeTrue())
 			})
 		})
 	})