+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG
 
 ## master
+* feature: filter gateway dataplanes through api and through `kumactl inspect dataplanes --gateway`
+  [#654](https://github.com/Kong/kuma/pull/654)
+* fix: added shorthand command name for mesh in kumactl
+  [#664](https://github.com/Kong/kuma/pull/664)
 * fix: make Envoy 1.13.1 work on Kind
   [#660](https://github.com/Kong/kuma/pull/660)
 * feat: added a new `kumactl install tracing` CLI command
@@ -31,8 +35,6 @@
   [#630](https://github.com/Kong/kuma/pull/630)
 * feature: added flag `--dry-run` for `kumactl apply`
   [#622](https://github.com/Kong/kuma/pull/622)
-* fix: added shorthand command name for mesh in kumactl
-  [#664](https://github.com/Kong/kuma/pull/664)
 
 ## [0.4.0]